Migrate subcription token APIs + Rename fcm file
