+++ .gitignore
@@ -2,6 +2,7 @@
 # Edit at https://www.toptal.com/developers/gitignore?templates=visualstudiocode,eclipse,java
 
 gcp-key.json
+fcm-key.json
 
 ### Eclipse ###
 .metadata
