+++ counsel-service/src/main/resources/application.properties
@@ -11,4 +11,4 @@ gcp.serviceAccountKeyPath=../gcp-key.json
 gcp.bucketName=hcmus-alumverse
 gcp.domainName=https://storage.googleapis.com/
 timeZone=Asia/Ho_Chi_Minh
-firebase.service.account.path=../service-account-key.json
+firebase.service.account.path=../fcm-key.json
