+++ group-service/src/main/resources/application.properties
@@ -10,4 +10,4 @@ spring.servlet.multipart.max-request-size = 25MB
 gcp.serviceAccountKeyPath=../gcp-key.json
 gcp.bucketName=hcmus-alumverse
 gcp.domainName=https://storage.googleapis.com/
-firebase.service.account.path=../service-account-key.json
+firebase.service.account.path=../fcm-key.json
