+++ notification-service/src/main/java/hcmus/alumni/notification/config/GlobalExceptionHandler.java
@@ -38,7 +38,7 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(MissingServletRequestParameterException.class)
     public ResponseEntity<ErrorResponse> handleMissingParams(MissingServletRequestParameterException ex) {
         String name = ex.getParameterName();
-        ErrorResponse error = new ErrorResponse(80000, "Tham số " + name + " bị thiếu");
+        ErrorResponse error = new ErrorResponse(100000, "Tham số " + name + " bị thiếu");
 
         return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
     }
@@ -46,14 +46,14 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(MissingServletRequestPartException.class)
     public ResponseEntity<ErrorResponse> handleMissingParts(MissingServletRequestPartException ex) {
         String name = ex.getRequestPartName();
-        ErrorResponse error = new ErrorResponse(80001, "File " + name + " bị thiếu");
+        ErrorResponse error = new ErrorResponse(100001, "File " + name + " bị thiếu");
 
         return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
     }
 
     @ExceptionHandler(MaxUploadSizeExceededException.class)
     public ResponseEntity<ErrorResponse> handleMaxSizeException(MaxUploadSizeExceededException ex) {
-        ErrorResponse error = new ErrorResponse(80002, "Dung lượng file không được vượt quá " + maxFileSize);
+        ErrorResponse error = new ErrorResponse(100002, "Dung lượng file không được vượt quá " + maxFileSize);
 
         return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
     }
@@ -62,7 +62,7 @@ public class GlobalExceptionHandler {
     public ResponseEntity<ErrorResponse> handleTypeMismatchParams(MethodArgumentTypeMismatchException ex) {
         String name = ex.getName();
         String requiredType = ex.getRequiredType().getSimpleName();
-        ErrorResponse error = new ErrorResponse(80003,
+        ErrorResponse error = new ErrorResponse(100003,
                 "Tham số " + name + " phải là " + requiredType);
 
         return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
@@ -70,7 +70,7 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(HttpMessageNotReadableException.class)
     public ResponseEntity<ErrorResponse> handleHttpMessageNotReadableException(HttpMessageNotReadableException ex) {
-        ErrorResponse error = new ErrorResponse(80004, "Request body không được để trống");
+        ErrorResponse error = new ErrorResponse(100004, "Request body không được để trống");
 
         return new ResponseEntity<>(error, HttpStatus.BAD_REQUEST);
     }
