+++ notification-service/src/main/java/hcmus/alumni/notification/model/user/UserSubscriptionTokenModel.java
@@ -1,12 +1,12 @@
-package hcmus.alumni.userservice.model;
+package hcmus.alumni.notification.model.user;
 
 import java.io.Serializable;
 import java.util.UUID;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
