+++ notification-service/src/main/java/hcmus/alumni/notification/repository/UserSubscriptionTokenRepository.java
@@ -1,4 +1,4 @@
-package hcmus.alumni.userservice.repository;
+package hcmus.alumni.notification.repository;
 
 import java.util.List;
 
@@ -8,8 +8,8 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 
-import hcmus.alumni.userservice.dto.IUserSubscriptionTokenDto;
-import hcmus.alumni.userservice.model.UserSubscriptionTokenModel;
+import hcmus.alumni.notification.dto.IUserSubscriptionTokenDto;
+import hcmus.alumni.notification.model.user.UserSubscriptionTokenModel;
 
 public interface UserSubscriptionTokenRepository extends JpaRepository<UserSubscriptionTokenModel, String> {
 	@Query("SELECT token FROM UserSubscriptionTokenModel token WHERE token.userId = :userId AND token.isDelete = false")
@@ -17,6 +17,6 @@ public interface UserSubscriptionTokenRepository extends JpaRepository<UserSubsc
 	
 	@Transactional
 	@Modifying
-	@Query(value = "UPDATE UserSubscriptionTokenModel SET isDelete = 1 WHERE id = :id")
+	@Query(value = "UPDATE UserSubscriptionTokenModel SET isDelete = true WHERE id = :id")
 	int deleteUserSubscriptionTokenById(String id);
 }
